(GOSVC-707) Support a ClientName field on ConnectionConfiguration.
The client name is combined with the first 8 characters of a sha1sum of
the connections originating ip:port (our old client identifier).

A client connecting with the name MyClientName will have a client
identifier of MyClientName-SHA.

If a client does not send a ClientName then we use the ip:port in place
of the ClientName (eg. 10.0.0.76:40334-SHA).

This new client identifier is used for logging and keeping track of
connections. Client level metrics also use this as the ClientIdentifier
label.

Change-Id: I9cf0abbd9a0f04c22fe3a2703a9a758e26446f57