fix: reinitialize activeMessages map on provider connect (NGMTS-25962)
Reinitialize the activeMessages map for the client whenever they connect
to the broker. Not emptying the map can result in thousands of messages
in the activeMessages map that will never be processed.

Messages are dependent on the rabbit channels they were consumed on so
these messages can never be consumed on a new connection or channel and
should be thrown away.

Change-Id: I2c29983437e7e4467f491c4784b70ebeb1e01760