fix: remove the complicated and not generally useful resubscribe logic (NGMTS-34079)
We attempt to resubscribe failed provider.Subscribe calls oh behalf of
the connected client, but if the client also resubscribes when getting
a ConsumedResponse with an error, we will end up with possibly dozens of
channels from arke the amqp091 broker.

This can happen on a slow consumer that's taking too long to process
messages. The channel to the broker will close, causing arke to
resubscribe. Ack's on in flight messages will fail, causing the client
to resubscribe and suddenly we have more channels open to the broker
than we should with some of the channels essentially being abandoned
because no client is listening to the streams (if the client did not
close the stream).

Change-Id: I5bb696a4c632ceb859d5370039fde03b7599b6c6