fix: event delivery delay (PSGO-530) (#42)
* fix: event delivery delay (PSGO-530)

Create a Cancel context from the grpc context. When Consume exits
we cancel the context. All of the long running goroutines for this
Consume will listen for Done() and shutdown.

Change our exchange channels to buffered from unbuffered.

Update the channel that sends messages directly from rabbit to be
processed to a buffered channel.