fix: prevent shutdown panic (PSGO-2041) (#150)
* fix: prevent shut down panic (PSGO-2041)

The arke subscriber code has a race condition that is caused by clients
disconnecting. When a client subscribes we create three goroutines. One
is the parent goroutine that manages the state of the client connection,
another goroutine manages the flow of messages to the client and the third
connects to the broker and manages the flow of messages from the broker.
Because of the parent/child relationship between goroutines and the nature
of both sides of the proxy can terminate the subscribe we have a two-way
shut down in place. The child goroutines use a go channel to indicate to
the parent that it needs to shut down. When the client indicates a shutdown,
it is possible the parent goroutine closes the channel before the child
goroutines are shutdown. If that happens the child goroutines may send on
the closed channel indicating that it is shutting down. Sending on a closed
channel causes a panic.

The two way shut down is tricky, the best solution is to recover the panic
to prevent it from crashing Arke.